\name{projektMOW-package}
\alias{projektMOW-package}
\alias{projektMOW}
\docType{package}
\title{
\packageTitle{projektMOW}
}
\description{
\packageDescription{projektMOW}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{projektMOW}
\packageIndices{projektMOW}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{projektMOW}

Maintainer: \packageMaintainer{projektMOW}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
