% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projekt.R
\name{tournament}
\alias{tournament}
\title{\code{tournament}Tournament of individuals}
\usage{
tournament(it, df)
}
\arguments{
\item{it}{iterator}

\item{df}{dataframe}
}
\value{
winner of tournament
}
\description{
\code{tournament}Tournament of individuals
}
